+++ keanu-project/src/main/java/io/improbable/keanu/vertices/dbl/Differentiable.java
@@ -1,4 +1,4 @@
-package io.improbable.keanu.vertices.dbl.probabilistic;
+package io.improbable.keanu.vertices.dbl;
 
 import java.util.List;
 import java.util.Map;
